add LoginPageClient because of an error: useSearchParams() should be wrapped in a suspense boundary at page '/login'
